refactor: main 코드와 같은 디렉터리로 이동
