+++ src/test/java/wooteco/subway/dao/memory/StationMemoryDaoTest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dao;
+package wooteco.subway.dao.memory;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -11,7 +11,7 @@ import wooteco.subway.dao.memory.StationMemoryDao;
 import wooteco.subway.domain.Station;
 import wooteco.subway.exception.NoSuchStationException;
 
-class StationDaoTest {
+class StationMemoryDaoTest {
 
     @BeforeEach
     void setUp() {
